Add your first controller
Controllers are what connect your business logic and database to the
templates that your app renders for the browser.

* **M**odel
  database and business logic

* **V**iew
  templates rendered for the browser or API client

* **C**ontroller
  accepts user input and/or API client calls and converts to commands
  for the model

It is generally best to think "Fat model, skinny controller" --
following this mantra will make your app easier to refactor and build a
better foundation for an API should you choose to add one later.

Logic in your controller is generally one-off, whereas logic in your
model can be reused by any part of your app.

Let's create a controller now:

    ⚡ rails g controller posts

Edit your controller so it does something:

    class PostsController < ApplicationController
      def index
        @posts = Post.all
        render json: @posts
      end
    end

Now visit your controller in the browser: http://localhost:3000/posts

Whoops! We got a routing error:

    Routing Error

    No route matches [GET] "/posts"

In the next step we'll add a route so the Rails app knows what to do
when we request `/posts`.