Set the default timezone
Rails is great at handling timezones for you and makes it easy to have
different timezone settings for each customer, user, or whatever.

For now, let's just tell Rails that all our users are in a single
timezone... Edit the `config/application.rb` file and uncomment the
following line:

    config.time_zone = 'Central Time (US & Canada)'

Now, this change won't take effect until you restart Rails. It's just a
quirk of Rails -- things in `config` generally don't take effect without
a restart of the web server.

`Ctrl-C` the web server process and start it again with:

    ⚡ rails server

Now refresh your browser and notice the timestamps on posts are right!