Add an action to create posts
Let's create an action in the PostsController:

    def create
      @post = Post.create(post_params)
      redirect_to action: :index
    end

Try to post the form and you'll get an error. `post_params` isn't
defined yet. Let's create that:

    private

    def post_params
      params.require(:post).permit(:body)
    end

This method returns the params submitted by the form that are relevant
to creating a new post. Some explanation of how Rails handles URL and
and post body params is in order...

When you used the Rails form helpers to generate the form, the `body`
param was named `post[body]`. When the form is submitted, all the params
wrapped in `post[*]` are grouped together in a Ruby hash.

The `post_params` method we created first grabs all the params with the
`post` prefix. This is the `require(:post)` bit above. If none exist,
then Rails throws an error.

Next, `post_params` filters out all `post`-prefixed params except for
`body`. It does this for safety. We don't want the user to modify the
timestamp or other attributes they shouldn't be able to change.

In summary, here's what goes into this method:

    {
      post: {
        body: 'Hello from the web',
        naughty: 'something malicious'
      }
    }

...and what is returned by this method:

    {
      body: 'Hello from the web'
    }