Show error messages
Rails has something called the "flash". The flash is a storage area tied
to the current session which allows a controller action to persist a
message to be displayed on the subsequent request. This is generally
used to show a message to the user after being redirected.

Devise sets a flash message when the user enters an incorrect password,
and we aren't showing the flash message, so the user is just presented
with the login form again. Let's fix that!

Add the following markup to your application layout:

    <% flash.each do |type, message| %>
      <div class="alert alert-warning" role="alert"><%= message %></div>
    <% end %>

This code simply shows anything in the flash as a warning. If you want,
you could change the style of the alert box based on the `type` of the
flash message (e.g. success, warning, danger, etc.).

Try logging in with the wrong password again, and you should see an
error message.

While we're improving the looks of things, let's add some margin to our
`nav` element. Add CSS to `app/assets/stylesheets/application.css`:

    nav {
      margin: 5px 0 15px;
    }