Tie each post to a user
It would be nice to see who posted what, and then only to allow the
author of a post to delete their post. This is very simple with Rails.

First, let's create a migration:

    ⚡ rails g migration add_user_id_to_posts

Now edit the new migration file in `db/migrations`:

    class AddUserIdToPosts < ActiveRecord::Migration
      def change
        change_table :posts do |t|
          t.integer :user_id
        end
      end
    end

Here we're telling Rails to modify our `posts` table so it has a new
`user_id` INT column.

Run the migration so the table is changed:

    ⚡ rake db:migrate

Now, modify the models so that the User model knows about Post, and vice
versa:

    # app/models/post.rb
    class Post < ActiveRecord::Base
      belongs_to :user
    end

    # app/models/user.rb
    class User < ActiveRecord::Base
      # ... snip ...
      has_many :posts
    end

Last, update the controller so it connects posts with the
`current_user`:

    class PostsController
      # ... snip ...

      def create
        @post = current_user.posts.create(post_params)
        redirect_to action: :index
      end

      def destroy
        @post = current_user.posts.find(params[:id])
        @post.destroy
        redirect_to action: :index
      end
    end

See if it works! You should only be able to create a post if you're
signed in, and you should only be able to delete posts you own.

Next, we'll hide some of the view elements dependending on the user
being signed in.