Hide UI elements depending on the signed in user
If the user is not signed in, then we don't want them to be able to make
a new post. Let's wrap the post form in a conditional:

    <% if user_signed_in? %>
      <!-- form markup -->
    <% end %>

Next, let's hide the delete button for posts the user doesn't own:

    <% if post.user == current_user %>
      <!-- delete button markup -->
    <% end %>

That's it! Play around with different accounts and check the page while
being signed out.